Fix up traceback logging
Trying to diagnose problems in plugins and such is near impossible
as-is, the exception logging system rarely prints anything of use. So
I've just ripped it out. It's a bit verbose, but too much verbosity is
better than no information at all.

Since it no longer does anything, the --trace switch has been removed.
This causes most of the tests to fail, so they've been updated
accordingly.